+++ src/internal/connector/exchange/service_iterators_test.go
@@ -653,6 +653,7 @@ func (suite *ServiceIteratorsSuite) TestFilterContainersAndFillCollections_Dupli
 						sc.scope,
 						test.inputMetadata(t, sc.cat),
 						control.Options{FailureHandling: control.FailFast},
+						true,
 						fault.New(true))
 					require.NoError(t, err, "getting collections", clues.ToCore(err))
 
@@ -902,6 +903,7 @@ func (suite *ServiceIteratorsSuite) TestFilterContainersAndFillCollections_Dupli
 				scope,
 				test.inputMetadata,
 				control.Options{FailureHandling: control.FailFast},
+				true,
 				fault.New(true))
 			require.NoError(t, err, "getting collections", clues.ToCore(err))